Minor Changes | Module Renames
Renamed some modules which were mistakenly named in caml case instead of snake.
Added a function inside of cluster config module to generate a cluster connection object.
